pyproject.toml
[project]
name = "iris-classifier"
version = "0.1.0"
description = "Iris flower classification API"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "scikit-learn",
    "pandas",
    "numpy",
    "joblib",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

